refactor(cost-tracking): merge diff_seed_pricing into reconcile_models
The two scripts shared llm-stats auth, seed I/O, and the same daily cadence
but lived in separate files with their own test suites. Folded
diff_seed_pricing into the renamed fetch_model_updates -> reconcile_models
so one invocation produces three signals from one upstream pass:

- new_models     - candidates not yet registered in OCS
- price_changes  - existing seed entries whose upstream rate has moved
                   (still rewrites llm_pricing.json + emits a rate-update
                   migration when non-empty)
- missing_pricing - OCS-managed (default_models) models with no llm_input
                    or llm_output seed entry. Replaces the digest signal
                    with a static check; surfaced via a GitHub issue from
                    the workflow.

GITHUB_OUTPUT now carries three groups of gates:
  has_new_models / new_model_count / new_model_ids
  has_price_changes / price_change_count / pricing_pr_title / pricing_pr_body_path
  has_missing_pricing / missing_pricing_count / missing_pricing_issue_body_path

Tests merged into one file. 7 new tests cover the missing-pricing audit
(active-vs-deleted distinction, required-kinds contract, ordering, etc.).
Workflow YAML refactor lands in the next commit.
